+++ src/Deprecated/Deprecated.php
@@ -25,6 +25,7 @@ class Deprecated {
             'business-profile-data-toll-free-number'          => 'toll_free_number',
             'business-profile-data-company-description'       => 'description',
             'business-profile-data-company-short-description' => 'short_description',
+            'business-profile-data-booking-link'              => 'booking_link',
 
         // This shortcode supports Business Profile Render elements
 
@@ -38,6 +39,7 @@ class Deprecated {
             'business-profile-render-toll-free-number'          => 'toll_free_number',
             'business-profile-render-company-description'       => 'description',
             'business-profile-render-company-short-description' => 'short_description',
+            'business-profile-render-booking-link'              => 'booking_link',
 
         ];